route-pattern: refactor type utils
move out of `src/lib/` and remove unused `Pretty` type
