+++ packages/route-pattern/src/lib/params.test.ts
@@ -1,5 +1,5 @@
 import type { Params } from './params.ts'
-import type { Assert, IsEqual } from './type-utils.d.ts'
+import type { Assert, IsEqual } from '../type-utils.d.ts'
 
 // prettier-ignore
 export type Tests = [
