+++ packages/route-pattern/src/lib/variant.test.ts
@@ -1,5 +1,5 @@
 import type { Variant } from './variant.ts'
-import type { Assert, IsEqual } from './type-utils.d.ts'
+import type { Assert, IsEqual } from '../type-utils.d.ts'
 
 // prettier-ignore
 export type Tests = [
